requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlmodel==0.0.14
sqlalchemy==2.0.23
pymysql==1.1.0
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Science and ML
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
plotly==5.17.0
matplotlib==3.10.3

# Reinforcement Learning
stable-baselines3==2.1.0
gymnasium==0.29.1
tensorboard==2.15.1

# Time Series Analysis
statsmodels==0.14.0

# Model Optimization
optuna==3.4.0
joblib==1.5.1

# LLM and AI Services
openai==1.3.7
langchain==0.1.0
langchain-openai==0.0.2

# Data Validation
pydantic==2.5.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0
aiofiles==23.2.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Azure and Cloud
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Monitoring and Logging
structlog==23.2.0 